Add patches to bring back the multiversion badge
This is what inspired this repository, as mentioned in the README. These
two patches bring back back the multi-version badge, either on the left
or the right of the poster card.

The canonical version is left, however, I could have sworn that earlier
last year it was actually on the right and it was changed to be on the
left. A change, that I begrudgingly accepted. Now that the badge is gone
altogether, I don't have to accept it because I just made a patch to
bring it back to begin with!