Add hovered preview thumbnail poster sample image
Shows the fact that when on the right, the multi-version-badge coexists
with the multi-select radio thumbnail. Which is [one of](https://forums.plex.tv/t/restore-duplicates-badge-removed-in-plex-web-4-53-0/711651/2)
the reasons why that the badge was removed...which is weird, because
putting it on the right is a simpler and 100x better option!